refactor(frontend): decompose AdminCostsPage into focused modules
Split 1,469-line monolith into 10 focused files under pages/admin/costs/:
- types.ts: all interfaces (~100 lines)
- constants.ts: COLORS, PROVIDER_COLORS, TIER_COLORS, PIE_COLORS, tooltipStyle
- formatters.ts: formatUSD, formatUSDPrecise, formatDate, formatMs
- shared.tsx: KPICard, ChartCard reusable components
- UserRequestsPanel.tsx: per-user request detail with pagination (~250 lines)
- UsersCostSection.tsx: users cost summary table with search (~175 lines)
- TransactionsSection.tsx: transactions table with filters + pagination (~170 lines)
- CohortSection.tsx: cohort analysis table (~95 lines)
- index.tsx: orchestrator with header, KPI cards, charts, model/tool tables

Original file reduced to 3-line re-export shim for backward compatibility.
